requirements.txt
discord.py>=2.3.2
motor>=3.3.2
Pillow>=10.2.0
requests>=2.31.0
dnspython>=2.6.1
python-dotenv>=1.0.1